fix: move ErrorDialog to composed and nit changes
